+++ app/Models/Seller.php
@@ -8,6 +8,7 @@ use ApiPlatform\Metadata\GetCollection;
 use ApiPlatform\Metadata\Post;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
 
 #[ApiResource(
     operations: [
@@ -32,5 +33,15 @@ class Seller extends Model
     public function products(): BelongsToMany
     {
         return $this->belongsToMany(Product::class, 'product_seller');
+    }
+
+    /**
+     * Define the relationship with the user.
+     *
+     * @return BelongsTo
+     */
+    public function user(): BelongsTo
+    {
+        return $this->belongsTo(User::class, 'user_id');
     }
 }