+++ app/Repositories/CityRepositoryInterface.php
@@ -9,4 +9,5 @@ interface CityRepositoryInterface
 {
     public function getCities(): Collection;
     public function getCityById(int $id): City;
+    public function getCityFromCoordinates(float $longitude, float $latitude): string;
 }